+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+### 2.3.3 (Apr 2018)
+* Updated: code improvements for using server-side options that DataTables provides through datatable-api plugin
+
 ### 2.3.2 (Apr 2018)
 * Added option to accept the privacy policy when user is not logged-in, by using an invite URL.
-* Small improvements in admin area.
+* Small improvements in admin section.
 
 ### 2.3.1 (Mar 2018)
 * Fixed: Check if user accepted If privacy notice is enabled on registration.